move search api call to backend
We're just doing the search from our own backend now. Because CORS is a
pain in the ass and the whole damn thing adds complexity that I don't
need. So now we'll do the search from the backend and just pass the
results on to the client; it's not as performant, but it's less prone to
breakage.